+++ java/geneticmlpcar/src/main/java/com/yzumis/geneticmlpcar/objects/MovableObject.java
@@ -5,15 +5,15 @@ package com.yzumis.geneticmlpcar.objects;
  */
 public abstract class MovableObject {
 
-    protected float x;
-    protected float y;
-    protected final float xSize;
-    protected final float ySize;
-    protected float xVelocity;
-    protected float yVelocity;
+    protected double x;
+    protected double y;
+    protected final double xSize;
+    protected final double ySize;
+    protected double xVelocity;
+    protected double yVelocity;
     private long lastUpdateTime;
 
-    public MovableObject(final float x, final float y, final float xSize, final float ySize, final float xVelocity) {
+    public MovableObject(final double x, final double y, final double xSize, final double ySize, final double xVelocity) {
         this.x = x;
         this.y = y;
         this.xSize = xSize;
@@ -24,7 +24,7 @@ public abstract class MovableObject {
 
     public void updatePosition(final long currentTimeMillis) {
         final long timeDifferenceMillis = currentTimeMillis - this.lastUpdateTime;
-        final float timeDifferenceSeconds = timeDifferenceMillis / 1000f;
+        final double timeDifferenceSeconds = timeDifferenceMillis / 1000d;
         this.lastUpdateTime = currentTimeMillis;
         this.x = this.x + timeDifferenceSeconds * this.xVelocity * 30;
         this.y = this.y + timeDifferenceSeconds * this.yVelocity * 60;
