+++ libraries/genetic/src/main/java/com/yzumis/genetic/IndividualFactory.java
@@ -2,6 +2,6 @@ package com.yzumis.genetic;
 
 public interface IndividualFactory {
 
-    <G> Individual createIndividual(final float mutationRate, final G goal);
+    <G> Individual createIndividual(final double mutationRate, final G goal);
 
 }
